refactor(shared): move TuiAgentConfig types out of the agent table
Why: tui-agent-config.ts sat at 297 of the 300 counted lines oxlint allows for
`src/**/*.ts`, so registering any further agent trips max-lines before the entry
is even written. The per-agent table now has 27 lines of headroom.

Type-only move — no behavior change. Consumers import from the new module
directly rather than through a re-export, per #14447.
